Extract helper method from MigrateProcedure
